Fixes Dafaflow Parquet issue by using finishBundle (#1047)

+++ pipelines/common/src/main/java/com/google/fhir/analytics/ParquetUtil.java
@@ -52,6 +52,11 @@ public class ParquetUtil {
   private final Timer timer;
   private final Random random;
   private final String namePrefix;
+  private boolean flushedInCurrentPeriod;
+
+  private synchronized void setFlushedInCurrentPeriod(boolean value) {
+    flushedInCurrentPeriod = value;
+  }
 
   public String getParquetPath() {
     return dwhFiles.getRoot();
@@ -90,6 +95,7 @@ public class ParquetUtil {
     this.writerMap = new HashMap<>();
     this.rowGroupSize = rowGroupSize;
     this.namePrefix = namePrefix;
+    setFlushedInCurrentPeriod(false);
     if (secondsToFlush > 0) {
       TimerTask task =
           new TimerTask() {
@@ -97,9 +103,14 @@ public class ParquetUtil {
             @Override
             public void run() {
               try {
-                log.info(
-                    "Flushing all Parquet writers for thread " + Thread.currentThread().getId());
-                flushAll();
+                // If a flush has happened recently, e.g., through a `FinishBundle` call, we don't
+                // need to do that again; the timer here is just to make sure the data is flushed
+                // into Parquet files _at least_ once in every `secondsToFlush`.
+                if (!flushedInCurrentPeriod) {
+                  log.info("Flush timed out for thread " + Thread.currentThread().getId());
+                  flushAll();
+                }
+                setFlushedInCurrentPeriod(false);
               } catch (IOException | ProfileException e) {
                 log.error("Could not flush Parquet files: " + e);
               }
@@ -174,10 +185,12 @@ public class ParquetUtil {
     }
   }
 
-  private synchronized void flushAll() throws IOException, ProfileException {
+  synchronized void flushAll() throws IOException, ProfileException {
+    log.info("Flushing all Parquet writers for thread " + Thread.currentThread().getId());
     for (String resourceType : writerMap.keySet()) {
       flush(resourceType);
     }
+    setFlushedInCurrentPeriod(true);
   }
 
   public synchronized void closeAllWriters() throws IOException {
